split service funcitionality in transformation service
